% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_benvo.R
\name{add_BEF}
\alias{add_BEF}
\title{Add Built Environment Feature to Benvo}
\usage{
add_BEF(x, bef_data, bef_id, d_function = sf::st_distance)
}
\arguments{
\item{x}{benvo or base benvo}

\item{bef_data}{'tidy' data frame containing date/spatial information for one unique bef}

\item{bef_id}{unique bef_id column name}

\item{d_function}{function for calculating distance. Default is \code{\link[sf]{st_distance}}}
}
\description{
Add Built Environment Feature to Benvo
}
