% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{aggrenvo}
\alias{aggrenvo}
\alias{aggrenvo,Benvo-method}
\title{Aggregate Matrix to Subject or Subject - Measurement Level}
\usage{
aggrenvo(x, M, stap_term, component, NA_to_zero = F)

\S4method{aggrenvo}{Benvo}(x, M, stap_term, component, NA_to_zero = F)
}
\arguments{
\item{x}{benvo object}

\item{M}{matrix to aggregate}

\item{stap_term}{relevant stap term}

\item{component}{one of c("Distance","Time","Distance-Time") indicating which column(s) of the bef dataset should be returned}
}
\description{
Aggregate Matrix to Subject or Subject - Measurement Level
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: method
}}

