% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_benvo.R
\name{base_benvo}
\alias{base_benvo}
\title{Base Benvo}
\usage{
base_benvo(subject_data, by, ...)
}
\arguments{
\item{subject_data}{data.frame containing subject level covariates.}

\item{by}{optional key}

\item{...}{optional arguments for specifying date-time columns see \code{\link{set_datetime_cols}}}
}
\value{
a benvo with attribute base = TRUE
}
\description{
When building a benvo iteratively the base benvo allows you to start with no bef data constructed a priori and build up from
subject data that contains spatial and/or temporal raw data in the form of \code{\link[sf]{sf}} structures or \code{\link[lubridate]{Date}} columns.
}
