% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{bef_summary,Benvo-method}
\alias{bef_summary,Benvo-method}
\title{benvo BEF Summary Method}
\usage{
\S4method{bef_summary}{Benvo}(x)
}
\arguments{
\item{x}{a benvo object}
}
\description{
benvo BEF Summary Method
}
