% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{bef_summary}
\alias{bef_summary}
\alias{bef_summary,Benvo-method}
\title{Benvo BEF Summary Generic}
\usage{
bef_summary(x)

\S4method{bef_summary}{Benvo}(x)
}
\arguments{
\item{x}{a benvo object}
}
\description{
Benvo BEF Summary Generic

benvo BEF Summary Method
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: 
}}

\keyword{internal}
