% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Benvo-class}
\alias{Benvo-class}
\alias{Benvo}
\title{Benvo data objects}
\description{
A custom data structure for handling point pattern built environment data. 
Consists of two dataframes (1) subject and (2) subject - BEF with a joining id(s).
}
\section{Slots}{

\describe{
\item{\code{subject_data}}{data.frame containing subject level covariates.}

\item{\code{bef_data}}{data.frame containing subject - bef distance/time data}

\item{\code{longitudinal}}{boolean value indicating whether the Benvo holds longitudinal data}

\item{\code{bef_names}}{character vector containing the names of the BEFs}

\item{\code{components}}{character vector containing the Space and/or Time data measured for each BEF}
}}

