% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo-methods.R
\name{benvo-methods}
\alias{benvo-methods}
\alias{bef_names}
\alias{components}
\alias{component_lookup}
\alias{num_BEF}
\alias{is.benvo}
\title{Benvo Methods}
\usage{
bef_names(x)

components(x)

component_lookup(x, term)

num_BEF(x)

is.benvo(x)
}
\arguments{
\item{x}{a benvo object}

\item{term}{bef_name string}
}
\description{
Benvo Methods
}
