% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo.R
\name{benvo}
\alias{benvo}
\title{Create a benvo object}
\usage{
benvo(subject_data, bef_data, by = NULL)
}
\arguments{
\item{subject_data}{data.frame containing subject level covariates.}

\item{bef_data}{named list of BEF data frames}

\item{by}{optional key}
}
\description{
Create a benvo object
}
\details{
benvo is a constructor function which creates benvo objects.
In particular, note that the \code{benvo} function will explicitly check the data you provide,
to ensure benvo methods can be performed without error.
}
