% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{bwinvo,Benvo-method}
\alias{bwinvo,Benvo-method}
\title{Between - Within Construction}
\usage{
\S4method{bwinvo}{Benvo}(x, M)
}
\arguments{
\item{x}{benvo object}

\item{M}{matrix to construct between/within measures}
}
\description{
Between - Within Construction
}
