% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{bwinvo}
\alias{bwinvo}
\title{Between - Within Construction}
\usage{
bwinvo(x, M)
}
\arguments{
\item{x}{benvo object}

\item{M}{matrix to construct between/within measures}
}
\description{
Between - Within Construction
}
\keyword{internal}
