% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{component_lookup,Benvo-method}
\alias{component_lookup,Benvo-method}
\title{Component look up}
\usage{
\S4method{component_lookup}{Benvo}(x, bef_name)
}
\arguments{
\item{x}{benvo object}

\item{bef_name}{bef_name string}
}
\description{
returns the measure (Space/Time) that's associated with the input BEF
}
