% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{component_lookup}
\alias{component_lookup}
\alias{component_lookup,Benvo-method}
\title{BEF component look up}
\usage{
component_lookup(x, bef_name)

\S4method{component_lookup}{Benvo}(x, bef_name)
}
\arguments{
\item{x}{benvo object}

\item{bef_name}{bef_name string}
}
\description{
BEF component look up

Component look up
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: returns the measure (Distance/Time) that's associated with the input BEF
}}

\keyword{internal}
