% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_CA_benvo.R
\name{create_CA_benvo}
\alias{create_CA_benvo}
\title{Create California Benvo}
\usage{
create_CA_benvo()
}
\value{
a benvo with the Los Angeles data converted to sf objects.
}
\description{
This function exists primarily to save the hassle of having an sf object stored as an
R data object, as it introduces non-ascii characters into the package.
With this function, the appropriate benvo is returned.
}
\seealso{
The building benvos vignette
}
