% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_BEF.R
\name{drop_BEF}
\alias{drop_BEF}
\title{Drop Built Environment Feature from Benvo}
\usage{
drop_BEF(x)
}
\arguments{
\item{x}{benvo or base benvo}
}
\value{
benvo without the active bef data
}
\description{
Remove the active BEF data table and corresponding sub-bef data from the benvo
}
