% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{example_benvo}
\alias{example_benvo}
\alias{FFbenvo}
\title{Small benvo for use in \pkg{benvo} examples and vignettes.}
\format{
A benvo with 1000 subjects and nearby simulated FFRs
\describe{
\item{\code{FFR_subjects }}{see FFR_subjects dataset}
\item{\code{FFR_distances}}{see FFR_distances dataset}
}
}
\usage{
FFbenvo
}
\description{
Small benvo for use in \pkg{benvo} examples and vignettes.
}
\keyword{datasets}
