% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{FFR_distances}
\alias{FFR_distances}
\title{Small dataset for use in \pkg{benvo} examples and vignettes.}
\format{
A data frame with 9501 rows and 2 columns
\describe{
\item{ \code{id}}{ The subject unique identifier}
\item{\code{Distance}}{The simulated distance between a hypothetical subject and fast food restaurant.}
}
}
\usage{
FFR_distances
}
\description{
Small dataset for use in \pkg{benvo} examples and vignettes.
}
\keyword{datasets}
