% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{FFR_subjects}
\alias{FFR_subjects}
\title{Small dataset for use in \pkg{benvo} examples and vignettes.}
\format{
A data frame with 1000 rows and 3 columns
\describe{
\item{\code{id}}{The subject unique identifier}
\item{\code{sex}}{The measurement unique identifier}
\item{\code{BMI}}{The Built Environment Unique identifier}
}
}
\usage{
FFR_subjects
}
\description{
Small dataset for use in \pkg{benvo} examples and vignettes.
}
\keyword{datasets}
