% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{HFS_distances_times}
\alias{HFS_distances_times}
\title{Small dataset for use in \pkg{benvo} examples and vignettes.}
\format{
A data frame with 5709 rows and 3 columns
\describe{
\item{\code{id}}{ The subject unique identifier}
\item{\code{measurement}}{The subject repeat measurement id}
\item{\code{Distance}}{The simulated distance between a hypothetical subject and fast food restaurant.}
\item{\code{Time}}{The simulated time between a hypothetical subject and fast food restaurant.}
}
}
\usage{
HFS_distances_times
}
\description{
Small dataset for use in \pkg{benvo} examples and vignettes.
}
\keyword{datasets}
