% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo-methods.R
\name{is.longitudinal}
\alias{is.longitudinal}
\alias{is.longitudinal.benvo}
\title{Is a longitudinal benvo}
\usage{
is.longitudinal(object)

\method{is.longitudinal}{benvo}(object)
}
\description{
Is a longitudinal benvo
}
\section{Methods (by class)}{
\itemize{
\item \code{benvo}: returns true if longitudinal
}}

\keyword{internal}
