% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{joining_ID,Benvo-method}
\alias{joining_ID,Benvo-method}
\title{benvo joining ID}
\usage{
\S4method{joining_ID}{Benvo}(x)
}
\arguments{
\item{x}{benvo object}
}
\description{
benvo joining ID
}
