% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{joining_ID}
\alias{joining_ID}
\title{Return the joining ID of the benvo}
\usage{
joining_ID(x)
}
\arguments{
\item{x}{benvo object}
}
\description{
Return the joining ID of the benvo
}
\keyword{internal}
