% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{LA_restaurants}
\alias{LA_restaurants}
\title{California Public Schools Fitnessgram Data}
\format{
A dataframe with 8101 rows and 4 columns
\describe{
\item{\code{name}}{Restaurant Name}
\item{\code{osm_id}}{openstreetmap unique id}
\item{\code{Latitude}}{Self Explanatory}
\item{\code{Longitude}}{Self Explanatory}
}
}
\usage{
LA_restaurants
}
\description{
California Public Schools Fitnessgram Data
}
\details{
data downloaded from the openstreetmap overpass api, subset to include just those restaurants in Los Angeles.
}
\keyword{datasets}
