% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{LA_schools}
\alias{LA_schools}
\title{California Public Schools Fitnessgram Data}
\format{
A dataframe with 308 rows and 8 columns
\describe{
\item{\code{Perc5c}}{Proportion of Obese 5th Graders}
\item{\code{NoStud5}}{Number of 5th  Graders in the class}
\item{\code{Charter}}{Factor variable indicating whether or not school is a charter school or not}
\item{\code{cdscode}}{School identifier}
\item{\code{City}}{Self Explanatory}
\item{\code{County}}{Self Explanatory}
\item{\code{Latitude}}{Self Explanatory}
\item{\code{Longitude}}{Self Explanatory}
}
}
\usage{
LA_schools
}
\description{
California Public Schools Fitnessgram Data
}
\details{
data downloaded from the CA department of education website, subset to include just those schools in Los Angeles.
}
\keyword{datasets}
