% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{longitudinal_design,Benvo-method}
\alias{longitudinal_design,Benvo-method}
\title{Extract Longitudinal Design Matrices}
\usage{
\S4method{longitudinal_design}{Benvo}(x, formula, ...)
}
\arguments{
\item{x}{benvo object}

\item{formula}{similar to \code{\link[lme4]{glmer}}.}

\item{...}{other arguments passed to the model frame}
}
\description{
For use with \code{\link[lme4]{glmer}} type formulas/models
}
