% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{longitudinal_HFS}
\alias{longitudinal_HFS}
\title{Small benvo for use in \pkg{benvo} longitudinal examples and vignettes.}
\format{
A benvo with 1000 subjects and nearby simulated FFRs
\describe{
\item{\code{HFS_subjects }}{see HFS_subjects dataset}
\item{\code{HFS_subjects}}{see HFS_distances dataset}
}
}
\usage{
longitudinal_HFS
}
\description{
Small benvo for use in \pkg{benvo} longitudinal examples and vignettes.
}
\details{
A hypothetical example showing how exposure to Healthy Food Stores (HFS) over time may decrease BMI
}
\keyword{datasets}
