% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{num_BEF,Benvo-method}
\alias{num_BEF,Benvo-method}
\title{Number of Built Environment Features}
\usage{
\S4method{num_BEF}{Benvo}(x)
}
\arguments{
\item{x}{a benvo object}
}
\description{
Number of Built Environment Features
}
