% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{num_BEF}
\alias{num_BEF}
\alias{num_BEF,Benvo-method}
\title{Number of BEF data frames}
\usage{
num_BEF(x)

\S4method{num_BEF}{Benvo}(x)
}
\arguments{
\item{x}{a benvo object}
}
\description{
Number of BEF data frames

Number of Built Environment Features
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: return the number of bef data frames
}}

\keyword{internal}
