% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{plot,Benvo-method}
\alias{plot,Benvo-method}
\title{Benvo plots}
\usage{
\S4method{plot}{Benvo}(x, term = NULL, component = NULL, p = 0.95)
}
\arguments{
\item{x}{benvo object}

\item{term}{name of BEF to plot. If NULL plots the first component listed in the Benvo.}

\item{component}{one of c("Distance","Time") indicating which measure to use. Defaults to Distance if both measures are available, otherwise uses the only option.}

\item{p}{The probability of distances/times that should be included in interval}
}
\description{
Plots the sorted distribution intervals of distances and times across subjects
}
