% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.benvo}
\alias{plot.benvo}
\title{Benvo plots}
\usage{
\method{plot}{benvo}(x, plotfun = "pointrange", ...)
}
\arguments{
\item{x}{benvo object}

\item{plotfun}{one of c("pointrange","map")}

\item{...}{extra arguments for plotfun}
}
\description{
Variety of plotting functions for benvo objects
}
