% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map}
\alias{plot_map}
\title{Spatial Plot of benvo}
\usage{
plot_map(x, term = NULL)
}
\arguments{
\item{x}{benvo object}

\item{term}{BEF term}
}
\description{
Provides a plot of benvo subjects and (one) BEF's locations
}
