% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{plot_pointrange,Benvo-method}
\alias{plot_pointrange,Benvo-method}
\title{Pointrange plot}
\usage{
\S4method{plot_pointrange}{Benvo}(x, BEF)
}
\arguments{
\item{x}{benvo object}

\item{BEF}{BEF specification}
}
\description{
Pointrange plot
}
