% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{plot_pointrange}
\alias{plot_pointrange}
\title{Pointrange plot}
\usage{
plot_pointrange(x, BEF, component)
}
\arguments{
\item{x}{benvo object}

\item{BEF}{BEF specification}

\item{component}{one of c("Distance","Time") indicating which column(s) of the bef dataset should be returned}
}
\description{
Pointrange plot
}
