% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Temporal Plot of benvo}
\usage{
plot_timeline(x, ...)
}
\arguments{
\item{x}{benvo object}

\item{...}{currently ignored}
}
\description{
Provides a plot of benvo subjects temporal exposure over time.
}
