% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{pointrange_plot}
\alias{pointrange_plot}
\alias{pointrange_plot,Benvo-method}
\title{benvo plot method}
\usage{
pointrange_plot(x, pars)

\S4method{pointrange_plot}{Benvo}(x, pars)
}
\arguments{
\item{pars}{optional vector of 
BEF strings with which to subset printed summary}

\item{object}{benvo object}
}
\description{
benvo plot method
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: 
}}

