% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbenvo-package.R
\docType{package}
\name{rbenvo-package}
\alias{rbenvo-package}
\alias{rbenvo}
\title{The 'rbenvo' package.}
\description{
rbenvo is package that holds S4 class objects and methods for built environment data to ease the use of working with these data and improve interoperability with other packages.
}
