% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R, R/distinct.R, R/filter.R,
%   R/group_by.R, R/mutate.R, R/select.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{distinct}
\alias{filter}
\alias{top_n}
\alias{group_by}
\alias{ungroup}
\alias{mutate}
\alias{transmute}
\alias{mutate_all}
\alias{mutate_at}
\alias{n}
\alias{select}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr:reexports]{contains}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr:reexports]{ends_with}}, \code{\link[dplyr:reexports]{everything}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:reexports]{matches}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr:mutate_all]{mutate_at}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr:reexports]{num_range}}, \code{\link[dplyr:reexports]{one_of}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:reexports]{starts_with}}, \code{\link[dplyr]{top_n}}, \code{\link[dplyr:mutate]{transmute}}, \code{\link[dplyr:group_by]{ungroup}}}
}}

