% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo.R
\name{set_datetime_cols}
\alias{set_datetime_cols}
\title{Set DateTime Columns}
\usage{
set_datetime_cols(
  measurement_date = NULL,
  start_date_col = NULL,
  stop_date_col = NULL
)
}
\arguments{
\item{measurement_date}{column string for the date at which a subject was measured}

\item{start_date_col}{column string for the date at which a subject/bef moved to their corresponding location}

\item{stop_date_col}{column string for the date at which a subject/bef stopped having exposure at the corresponding location.}
}
\description{
When exposure time and lag exposure time need to be calculated
the measurement date, and start/stop date columns can be provided to the benvo
and base_benvo functions as optional arguments. Note that these columns will be converted to \code{\link[lubridate]{Date}}
objects if they're not already.
}
