% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,Benvo-method}
\alias{show,Benvo-method}
\title{benvo Print Method}
\usage{
\S4method{show}{Benvo}(object)
}
\arguments{
\item{object}{benvo object}
}
\description{
benvo Print Method
}
