% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{subject_design,Benvo-method}
\alias{subject_design,Benvo-method}
\title{Extract Subject Design Matrix}
\usage{
\S4method{subject_design}{Benvo}(formula, x, ...)
}
\arguments{
\item{formula}{similar to \code{\link[stats]{lm}}.}

\item{x}{benvo object}

\item{...}{other arguments passed to the model frame}
}
\description{
Extract Subject Design Matrix
}
