% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{subject_design}
\alias{subject_design}
\title{Subject Design Matrix}
\usage{
subject_design(x, formula, ...)
}
\arguments{
\item{x}{benvo object}

\item{formula}{similar to \code{\link[stats]{lm}}.}

\item{...}{other arguments passed to the model frame}
}
\description{
Subject Design Matrix
}
