% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{summary,Benvo-method}
\alias{summary,Benvo-method}
\title{benvo Summary Method}
\usage{
\S4method{summary}{Benvo}(x)
}
\arguments{
\item{x}{a benvo object}
}
\description{
benvo Summary Method
}
