% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary.R
\name{summary.benvo}
\alias{summary.benvo}
\title{benvo BEF Summary Generic}
\usage{
\method{summary}{benvo}(object, ...)
}
\arguments{
\item{object}{a benvo object}

\item{...}{ignored}
}
\description{
benvo BEF Summary Generic
}
