% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Benvo-class}
\alias{Benvo-class}
\alias{Benvo}
\title{Benvo data objects}
\description{
A custom data structure for handling point pattern built environment data.
Consists of two types of dataframes (1) subject and (2) subject - BEF with a joining id(s).
Whether using the \code{\link[rbenvo]{benvo}} constructor helper function or the raw
\code{\link[rbenvo]{Benvo}} class constructor, certain constraints are maintained within the class.
Namely,(1) there must be a bef name and component for each bef data frame.
(2) Both bef_data and subject data must inherit
from class data.frame. (3) Components must be one of c("Distance","Time","Distance-Time")
}
\section{Slots}{

\describe{
\item{\code{subject_data}}{data.frame containing subject level covariates.}

\item{\code{bef_data}}{list containing subject - bef distance/time data frames}

\item{\code{longitudinal}}{boolean value indicating whether the Benvo holds longitudinal data}

\item{\code{bef_names}}{character vector containing the names of the BEFs}

\item{\code{components}}{character vector containing the Distances and/or Time data measured for each BEF}

\item{\code{id}}{character vector of length 1 or 2 that determines how subjects are joined. Note that
for repeated measures data it is understood that id[1] is the subject id and id[2] is the measurement/visit id}
}}

