% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo.R
\name{benvo}
\alias{benvo}
\title{Create a benvo object}
\usage{
benvo(subject_data, bef_data, by = NULL)
}
\arguments{
\item{subject_data}{data.frame containing subject level covariates.}

\item{bef_data}{named list of BEF data frames}

\item{by}{optional key}
}
\description{
Create a benvo object
}
\details{
benvo is a helper constructor function which creates nicely formatted Benvo objects.
In particular, note that the \code{benvo} function will explicitly alter the data you provide, creating a new
numeric joining ID to enable easy aggregation and use with other methods. This alteration will not occur
if calling the raw constructor function \code{\link[rbenvo]{Benvo}}, though it will check for it.
}
\seealso{
\code{\link[rbenvo]{Benvo}}
}
