% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo.R
\name{benvo}
\alias{benvo}
\title{Create a benvo object}
\usage{
benvo(subject_data, bef_data, bef_names, joining_id, distance_col = NULL,
  exposed_time_col = NULL)
}
\arguments{
\item{subject_data}{data.frame containing subject level covariates.}

\item{bef_data}{list of BEF data frames}

\item{bef_names}{character vector of data frames}

\item{joining_id}{character vector containing strings in both subject and bef data frames used to join data}

\item{distance_col}{character vector containing strings
in bef_data frames where Distance measures are stored. NULL if there is no distance measure.}

\item{exposed_time_col}{character vector containing strings
in bef_data frames where Distance measures are stored. NULL if there is no time measure.}
}
\description{
Create a benvo object
}
