% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo.R
\name{benvo}
\alias{benvo}
\title{Create a benvo object}
\usage{
benvo(subject_data, sub_bef_data = NULL, by = NULL, ...)
}
\arguments{
\item{subject_data}{data.frame containing subject level covariates.}

\item{sub_bef_data}{named list of data frames that contain subject-bef relevant data.
NULL by default which returns a "base benvo" Which can be built upon/added to.}

\item{by}{optional key to link subject - sub_bef data. Will use the intersection of column names if not specified directly.}

\item{...}{optional arguments for specifying date-time columns see \code{\link{set_datetime_cols}}}
}
\value{
benvo object
}
\description{
Create a benvo object
}
\details{
benvo is a constructor function which creates benvo objects.
In particular, note that the \code{benvo} function will explicitly check the data you provide,
to ensure benvo methods can be performed without error.
}
\seealso{
\href{https://apeterson91.github.io/rbenvo/articles/Introduction.html}{Introductory} and more \href{https://apeterson91.github.io/rbenvo/articles/Building_Benvos.html}{Specialized} vignettes.
}
