% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{bwinvo}
\alias{bwinvo}
\alias{bwinvo,Benvo-method}
\title{Between - Within Construction Generic}
\usage{
bwinvo(x, M)

\S4method{bwinvo}{Benvo}(x, M)
}
\arguments{
\item{x}{benvo object}

\item{M}{matrix to construct between/within measures}
}
\description{
Between - Within Construction Generic

Between - Within Construction
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: between within decomposition of longitudinal matrix M according to benvo subject-bef data
}}

\keyword{internal}
