% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{bwinvo}
\alias{bwinvo}
\alias{bwinvo.benvo}
\title{Between - Within Decomposition}
\usage{
bwinvo(x, M)

\method{bwinvo}{benvo}(x, M)
}
\arguments{
\item{x}{benvo object}

\item{M}{matrix to construct between/within measures}
}
\description{
Between - Within Decomposition

Between - Within Construction
}
\section{Methods (by class)}{
\itemize{
\item \code{benvo}: between within decomposition of longitudinal matrix M according to benvo subject-bef data
}}

\keyword{internal}
