% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{joining_ID}
\alias{joining_ID}
\alias{joining_ID,Benvo-method}
\title{Return the joining ID of the benvo}
\usage{
joining_ID(x)

\S4method{joining_ID}{Benvo}(x)
}
\arguments{
\item{x}{benvo object}
}
\description{
Return the joining ID of the benvo

benvo joining ID
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: return appropriate benvo joining ID
}}

\keyword{internal}
