% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{joinvo}
\alias{joinvo}
\alias{joinvo,Benvo-method}
\title{Join BEF and subject data within a benvo}
\usage{
joinvo(x, bef_name, component = "Distance", tibble = F, NA_to_zero = F)

\S4method{joinvo}{Benvo}(x, bef_name, component = "Distance", tibble = F, NA_to_zero = F)
}
\arguments{
\item{x}{benvo object}

\item{bef_name}{string of bef data to join on in bef_data}

\item{component}{one of c("Distance","Time","Distance-Time") indicating which column(s) of the bef dataset should be returned}

\item{tibble}{boolean value of whether or not to return a data.frame or tibble}

\item{NA_to_zero}{replaces NA values with zeros - potentially useful when constructing design matrices}
}
\description{
Join BEF and subject data within a benvo
}
\details{
Joins the subject dataframe within a benvo to the supplied BEF dataframe keeping the selected component
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: method
}}

