% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{joinvo}
\alias{joinvo}
\alias{joinvo.benvo}
\title{Join BEF and subject data within a benvo}
\usage{
joinvo(x, term, component = "Distance", NA_to_zero = F)

\method{joinvo}{benvo}(x, term, component = "Distance", NA_to_zero = F)
}
\arguments{
\item{x}{benvo object}

\item{term}{string of bef name to join on in sub_bef_data}

\item{component}{one of c("Distance","Time","Distance-Time") indicating which column(s) of the bef dataset should be returned}

\item{NA_to_zero}{replaces NA values with zeros - potentially useful when constructing design matrices}
}
\description{
Join BEF and subject data within a benvo
}
\details{
Joins the subject dataframe within a benvo to the supplied BEF dataframe keeping the selected component
}
\section{Methods (by class)}{
\itemize{
\item \code{benvo}: method
}}

