% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{LA_restaurants}
\alias{LA_restaurants}
\alias{LA_FF}
\title{Los Angeles Fast Food Restaurants}
\format{
A dataframe with 8101 rows and 4 columns
\describe{
\item{\code{name}}{Restaurant Name}
\item{\code{osm_id}}{openstreetmap unique id}
\item{\code{Latitude}}{Self Explanatory}
\item{\code{Longitude}}{Self Explanatory}
}
}
\usage{
LA_FF
}
\description{
Los Angeles Fast Food Restaurants
}
\details{
data downloaded from the openstreetmap overpass api classified as "amenity:fast_food".
}
\keyword{datasets}
