% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{longitudinal_design}
\alias{longitudinal_design}
\alias{longitudinal_design,Benvo-method}
\title{Longitudinal design dataframe}
\usage{
longitudinal_design(x, formula, ...)

\S4method{longitudinal_design}{Benvo}(x, formula, ...)
}
\arguments{
\item{x}{benvo object}

\item{formula}{similar to \code{\link[lme4]{glmer}}.}

\item{...}{other arguments passed to the model frame}
}
\description{
For use with \code{\link[lme4]{glmer}} type formulas/models
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: method
}}

