% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{longitudinal_design}
\alias{longitudinal_design}
\title{Longitudinal design dataframe}
\usage{
longitudinal_design(x, formula, ...)

longitudinal_design(x, formula, ...)
}
\arguments{
\item{x}{benvo object}

\item{formula}{similar to \code{\link[lme4]{glmer}}.}

\item{...}{other arguments passed to the model frame}
}
\description{
For use with \code{\link[lme4]{glmer}} type formulas/models
}
\section{Functions}{
\itemize{
\item \code{longitudinal_design}: method
}}

