% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{plot_pointrange}
\alias{plot_pointrange}
\alias{plot_pointrange,Benvo-method}
\title{Pointrange plot}
\usage{
plot_pointrange(x, BEF, component, p = 0.95)

\S4method{plot_pointrange}{Benvo}(x, BEF, component, p = 0.95)
}
\arguments{
\item{x}{benvo object}

\item{BEF}{BEF specification}

\item{component}{one of c("Distance","Time") indicating which column(s) of the bef dataset should be returned}

\item{p}{The probability of distances/times that should be included in interval}
}
\description{
Plots the sorted distribution intervals of distances and times across subjects
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: method
}}

