% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{plot_pointrange}
\alias{plot_pointrange}
\alias{plot_pointrange,Benvo-method}
\title{Pointrange plot}
\usage{
plot_pointrange(x, BEF, component)

\S4method{plot_pointrange}{Benvo}(x, BEF, component)
}
\arguments{
\item{x}{benvo object}

\item{BEF}{BEF specification}

\item{component}{one of c("Distance","Time") indicating which column(s) of the bef dataset should be returned}
}
\description{
Pointrange plot

Pointrange plot
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: pointrange plot for bef data
}}

\keyword{internal}
