% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pointrange}
\alias{plot_pointrange}
\title{Plot Pointrange}
\usage{
plot_pointrange(x, term = NULL, component = NULL, p = 0.95)
}
\arguments{
\item{x}{benvo object}

\item{term}{name of BEF to plot. If NULL plots the first component listed in the Benvo.}

\item{component}{one of c("Distance","Time") indicating which measure to use. Defaults to Distance if both measures are available, otherwise uses the only option.}

\item{p}{The probability of distances/times that should be included in interval}
}
\description{
Plot Pointrange
}
