% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{subject_design}
\alias{subject_design}
\alias{subject_design,Benvo-method}
\title{Subject Design Matrix}
\usage{
subject_design(x, formula, ...)

\S4method{subject_design}{Benvo}(x, formula, ...)
}
\arguments{
\item{x}{benvo object}

\item{formula}{similar to \code{\link[stats]{lm}}.}

\item{...}{other arguments passed to the model frame}
}
\description{
Subject Design Matrix

Extract Subject Design Matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: subject design method
}}

\keyword{internal}
