% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{subject_design}
\alias{subject_design}
\title{Extract Subject Design Matrix}
\usage{
subject_design(formula, x, ...)
}
\arguments{
\item{formula}{similar to \code{\link[stats]{lm}}.}

\item{x}{benvo object}

\item{...}{other arguments passed to the model frame}
}
\description{
Extract Subject Design Matrix
}
