% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{subject_design}
\alias{subject_design}
\alias{subject_design.benvo}
\title{Subject Design Matrix}
\usage{
subject_design(x, formula, ...)

\method{subject_design}{benvo}(x, formula, ...)
}
\arguments{
\item{x}{benvo object}

\item{formula}{similar to \code{\link[stats]{lm}}.}

\item{...}{optional arguments}
}
\description{
Subject Design Matrix

Extract Subject Design Matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{benvo}: method
}}

\keyword{internal}
